[pf-702] rename getUserCampaignParticipations
To findUserCampaignParticipations
