+++ api/lib/domain/usecases/find-user-campaign-participations.js
