+++ api/lib/domain/usecases/index.js
@@ -63,6 +63,8 @@ module.exports = injectDependencies({
   createOrganization: require('./create-organization.js'),
   createSession: require('./create-session'),
   createUser: require('./create-user'),
+  findCampaignParticipationsByAssessment: require('./find-campaign-participations-by-assessment'),
+  findCampaignParticipationsWithResults: require('./find-campaign-participations-with-results'),
   findCertificationAssessments: require('./find-certification-assessments'),
   findCertificationCenters: require('./find-certification-centers'),
   findCompetenceEvaluations: require('./find-competence-evaluations'),
@@ -72,6 +74,7 @@ module.exports = injectDependencies({
   findSessionsForCertificationCenter: require('./find-sessions-for-certification-center'),
   findSmartPlacementAssessments: require('./find-smart-placement-assessments'),
   findSnapshots: require('./find-snapshots.js'),
+  findUserCampaignParticipations: require('./find-user-campaign-participations'),
   findUsers: require('./find-users.js'),
   getAssessment: require('./get-assessment'),
   getAttendanceSheet: require('./get-attendance-sheet'),
@@ -79,8 +82,6 @@ module.exports = injectDependencies({
   getCampaignByCode: require('./get-campaign-by-code'),
   getCampaignParticipation: require('./get-campaign-participation'),
   getCampaignParticipationResult: require('./get-campaign-participation-result'),
-  findCampaignParticipationsByAssessment: require('./find-campaign-participations-by-assessment'),
-  findCampaignParticipationsWithResults: require('./find-campaign-participations-with-results'),
   getCampaignReport: require('./get-campaign-report'),
   getCertificationCenter: require('./get-certification-center'),
   getCorrectionForAnswerWhenAssessmentEnded: require('./get-correction-for-answer-when-assessment-ended'),
@@ -99,7 +100,6 @@ module.exports = injectDependencies({
   getScorecard: require('./get-scorecard'),
   getProgression: require('./get-progression'),
   getUser: require('./get-user'),
-  getUserCampaignParticipations: require('./get-user-campaign-participations'),
   getUserCertification: require('./get-user-certification'),
   getUserCertificationWithResultTree: require('./get-user-certification-with-result-tree'),
   getUserCertificationCenterMemberships: require('./get-user-certification-center-memberships'),
